add rsync

+++ remote-backup/run.sh
@@ -12,6 +12,11 @@ REMOTE_DIRECTORY=$(jq --raw-output ".remote_directory" $CONFIG_PATH)
 ZIP_PASSWORD=$(jq --raw-output '.zip_password' $CONFIG_PATH)
 KEEP_LOCAL_BACKUP=$(jq --raw-output '.keep_local_backup' $CONFIG_PATH)
 
+RSYNC_HOST=$(jq --raw-output ".rsync_host" $CONFIG_PATH)
+RSYNC_ROOTFOLDER=$(jq --raw-output ".rsync_rootfolder" $CONFIG_PATH)
+RSYNC_USER=$(jq --raw-output ".rsync_user" $CONFIG_PATH)
+RSYNC_PASSWORD=$(jq --raw-output ".rsync_password" $CONFIG_PATH)
+
 # create variables
 SSH_ID="${HOME}/.ssh/id"
 
@@ -37,16 +42,17 @@ function add-ssh-key {
 
 function copy-backup-to-remote {
 
-    cd /backup/
-    if [[ -z $ZIP_PASSWORD  ]]; then
-      echo "Copying ${slug}.tar to ${REMOTE_DIRECTORY} on ${SSH_HOST} using SCP"
-      scp -F "${HOME}/.ssh/config" "${slug}.tar" remote:"${REMOTE_DIRECTORY}"
-    else
-      echo "Copying password-protected ${slug}.zip to ${REMOTE_DIRECTORY} on ${SSH_HOST} using SCP"
-      zip -P "$ZIP_PASSWORD" "${slug}.zip" "${slug}".tar
-      scp -F "${HOME}/.ssh/config" "${slug}.zip" remote:"${REMOTE_DIRECTORY}" && rm "${slug}.zip"
+    if hass.config.false 'ssh_enabled'; then
+        cd /backup/
+        if [[ -z $ZIP_PASSWORD  ]]; then
+            echo "Copying ${slug}.tar to ${REMOTE_DIRECTORY} on ${SSH_HOST} using SCP"
+            scp -F "${HOME}/.ssh/config" "${slug}.tar" remote:"${REMOTE_DIRECTORY}"
+        else
+            echo "Copying password-protected ${slug}.zip to ${REMOTE_DIRECTORY} on ${SSH_HOST} using SCP"
+            zip -P "$ZIP_PASSWORD" "${slug}.zip" "${slug}".tar
+            scp -F "${HOME}/.ssh/config" "${slug}.zip" remote:"${REMOTE_DIRECTORY}" && rm "${slug}.zip"
+        fi
     fi
-
 }
 
 function delete-local-backup {
@@ -78,6 +84,20 @@ function create-local-backup {
     echo "Creating local backup: \"${name}\""
     slug=$(hassio snapshots new --options name="${name}" | jq --raw-output '.data.slug')
     echo "Backup created: ${slug}"
+}
+
+function rsync_folders {
+
+    if hass.config.false 'rsync_enabled'; then
+        rsyncurl="$RSYNC_USER@$RSYNC_HOST::$RSYNC_ROOTFOLDER"
+        echo "[Info] trying to rsync hassio folders to $rsyncurl"
+         sshpass -p $RSYNC_PASSWORD rsync -av /config/ $rsyncurl/config/ 
+         sshpass -p $RSYNC_PASSWORD rsync -av /addons/ $rsyncurl/addons/ 
+         sshpass -p $RSYNC_PASSWORD rsync -av /backup/ $rsyncurl/backup/ 
+         sshpass -p $RSYNC_PASSWORD rsync -av /share/ $rsyncurl/share/ 
+         sshpass -p $RSYNC_PASSWORD rsync -av /ssl/ $rsyncurl/ssl/ 
+        echo "[Info] Finished rsync"
+    fi
 }
